requirements.txt
pyspark==3.5.1
pandas==2.2.2
pyyaml==6.0.2
python-dotenv==1.0.1
cryptography==42.0.8
pytest==8.3.2